requirements.txt
Appium-Python-Client==4.0.1
pytest==8.2.2
python-dotenv==1.0.1
requests==2.32.3
selenium==4.22.0



